Merged questions.py with registration.py and changed every file that depended on questions.py
