+++ website/registrations/models/__init__.py
@@ -1,3 +1,2 @@
 from registrations.models.employee import Employee
-from registrations.models.registration import Registration
-from registrations.models.questions import *
+from registrations.models.registration import *
